Compiles

+++ Snippets/01. gcd/01. gcd.c
@@ -1,15 +1,15 @@
 #include "./01. gcd.h"
 
-void swap(size_t *x, size_t *y){
-    size_t z;
+void swap(unsigned long long *x, unsigned long long *y){
+    unsigned long long z;
 
     z = *x;
     *x = *y;
     *y = z;
 }
 
-size_t gcd(size_t u, size_t v){
-    size_t t;
+unsigned long long gcd(unsigned long long u, unsigned long long v){
+    unsigned long long t;
 
     u = abs(u);
     v = abs(v);
@@ -22,8 +22,8 @@ size_t gcd(size_t u, size_t v){
     return v;
 }
 
-size_t gcd_mod(size_t u, size_t v){
-    size_t t;
+unsigned long long gcd_mod(unsigned long long u, unsigned long long v){
+    unsigned long long t;
 
     //u = abs(u);
     //v = abs(v);
@@ -38,12 +38,12 @@ size_t gcd_mod(size_t u, size_t v){
     return v;
 }
 
-size_t gcd_3(size_t x, size_t y, size_t z){
+unsigned long long gcd_3(unsigned long long x, unsigned long long y, unsigned long long z){
     return gcd_mod(gcd_mod(x,y), z);
 }
 
 void main_gcd(){
-    size_t x, y;
+    unsigned long long x, y;
     while(scanf("%d %d", &x, &y) != EOF)
         if(x>0 && y>0)
             printf("%d %d %d\n", x, y, gcd(x,y));
